+++ CHANGELOG.markdown
@@ -2,6 +2,7 @@
 
  * Updated Rakefile to run tests under ruby 1.8.7 and ruby 1.9.2
  * Added `page_size` parameter to `total_pages_in` to allow for checking what if values in that scenario
+ * Added `page_size` parameter to `leaders` and `around_me` calls
 
 # leaderboard 1.0.4 (2011-04-26)